migration_manager/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-spanner
requests